refactor: InProgressTaskItem refactoring
